+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+- [@react-enhanced/hooks](./packages/@react-enhanced/hooks/CHANGELOG.md)
 - [react-qrcode](./packages/react-qrcode/CHANGELOG.md)
 - [react-qrious](./packages/react-qrious/CHANGELOG.md)
 - [react-rx](./packages/react-rx/CHANGELOG.md)